Convert from AMD to ES6
This is a continuation of the previous commit. Most of the files were
converted using an npm amd-to-es6 conversion tool. I forget which one.
