+++ app/javascript/app/app-init.js
@@ -1,18 +1,21 @@
-// Manages the application initialization for most pages.
-// This script is called by the homepage, search result
-// and search details pages. It is not called by
-// the about page.
-require([
-  'util/translation/google-translate',
-  'app/popup/popups',
-  'app/alerts'
-],
-function (googleTranslate, popups, alerts) {
-  'use strict';
+// Manages the application initialization for all pages.
 
-  // If page is not translated, initialize the header popups.
-  if (!googleTranslate.isTranslated())
-    popups.init();
+import googleTranslate from 'app/util/translation/google-translate';
+import popups from 'app/popup/popups';
+import alerts from 'app/alerts';
 
-  alerts.init();
-});
+// The google-translate script handles loading of the
+// Google Website Translator Gadget at the bottom of the page's body.
+// The layout settings passed in as an argument to the initialization
+// method can be set to:
+// InlineLayout.VERTICAL, InlineLayout.HORIZONTAL,
+// which correspond to the 'inline' display modes available through Google.
+// The 'tabbed' and 'auto' display modes are not supported.
+// The 'inline' InlineLayout.SIMPLE layout is also not supported.
+googleTranslate.init(googleTranslate.InlineLayout.VERTICAL);
+
+// If page is not translated, initialize the header popups.
+if (!googleTranslate.isTranslated())
+  popups.init();
+
+alerts.init();
