--- app/javascript/app/routes/home/index.js
@@ -1,9 +1,0 @@
-// Manages initialization of scripts for the home page.
-require([
-  'app/app-init',
-  'home/home-init',
-  'application'
-],
-function () {
-  'use strict';
-});
