--- app/javascript/app/routes/locations/index.js
@@ -1,10 +1,0 @@
-// Manages initialization of scripts for the search results page.
-require([
-  'app/app-init',
-  'search/search-init',
-  'result/result-init',
-  'application'
-],
-function () {
-  'use strict';
-});
