--- app/javascript/app/routes/locations/show.js
@@ -1,10 +1,0 @@
-// Manages initialization of scripts for the search results page.
-require([
-  'app/app-init',
-  'search/search-init',
-  'detail/detail-init',
-  'application'
-],
-function () {
-  'use strict';
-});
