+++ app/javascript/app/search/filter/search-filters.js
@@ -1,56 +1,50 @@
 // Handles search filter functionality.
-define([
-  'search/filter/TextInput',
-  'util/geolocation/geolocate-action',
-  'domReady!'
-],
-function (TextInput, geo) {
-  'use strict';
+import TextInput from 'app/search/filter/TextInput';
+import geo from 'app/util/geolocation/geolocate-action';
 
-  // The search filters.
-  var _keyword;
-  var _location;
-  var _agency;
+// The search filters.
+var _keyword;
+var _location;
+var _agency;
 
-  // The form to submit.
-  var _searchForm;
+// The form to submit.
+var _searchForm;
 
-  // Main module initialization.
-  function init() {
+// Main module initialization.
+function init() {
 
-    // Set up geolocation button.
-    geo.init('button-geolocate', _geolocationClicked);
+  // Set up geolocation button.
+  geo.init('button-geolocate', _geolocationClicked);
 
-    // Set up text input filters
-    _keyword = TextInput.create('keyword-search-box');
-    _location = TextInput.create('location-options');
-    _agency = TextInput.create('org-name-options');
+  // Set up text input filters
+  _keyword = TextInput.create('keyword-search-box');
+  _location = TextInput.create('location-options');
+  _agency = TextInput.create('org-name-options');
 
-    // Capture form submission.
-    _searchForm = document.getElementById('form-search');
+  // Capture form submission.
+  _searchForm = document.getElementById('form-search');
 
-    // Hook reset button on the page and listen for a click event.
-    var resetButton = document.getElementById('button-reset');
-    resetButton.addEventListener('click', _resetClicked, false);
-  }
+  // Hook reset button on the page and listen for a click event.
+  var resetButton = document.getElementById('button-reset');
+  resetButton.addEventListener('click', _resetClicked, false);
+}
 
-  // The geolocation button was clicked in the location filter.
-  function _geolocationClicked(address) {
-    document.getElementById('location').value = address;
-    _searchForm.submit();
-  }
+// The geolocation button was clicked in the location filter.
+function _geolocationClicked(address) {
+  document.getElementById('location').value = address;
+  _searchForm.submit();
+}
 
-  // The clear filters button was clicked.
-  function _resetClicked(evt) {
-    _keyword.reset();
-    _location.reset();
-    _agency.reset();
+// The clear filters button was clicked.
+function _resetClicked(evt) {
+  _keyword.reset();
+  _location.reset();
+  _agency.reset();
 
-    evt.preventDefault();
-    evt.target.blur();
-  }
+  evt.preventDefault();
+  evt.target.blur();
+}
 
-  return {
-    init:init
-  };
-});
+export default {
+  init:init
+};
