--- app/javascript/app/search/search-init.js
@@ -1,9 +1,0 @@
-// Manages search initialization.
-require([
-  'search/filter/search-filters'
-],
-function (filters) {
-  'use strict';
-
-  filters.init();
-});
