+++ app/javascript/packs/about-index.js
@@ -1,6 +1,5 @@
 // Manages initialization of scripts for the About page.
 import FeedbackForm from 'app/FeedbackForm';
-require('application');
 
 // Initialize the feedback form.
 FeedbackForm.create('#feedback-box .feedback-form');
