+++ app/views/shared/_head.html.haml
@@ -35,7 +35,7 @@
 = stylesheet_link_tag 'https://fonts.googleapis.com/css?family=Quattrocento+Sans:400,400italic,700'
 = stylesheet_link_tag 'application', media: 'all'
 
--# Include third-party scripts
+-# Include third-party scripts.
 = javascript_include_tag 'vendor'
 
 -# Include Internet Explorer polyfills for legacy system compatibility.
